# backend/docs/swagger.yaml
basePath: /
definitions:
...
  database.Fishbone:
    properties:
      environment:
        type: string
      machine:
        type: string
      man:
        type: string
      material:
        type: string
      measurement:
        type: string
      method:
        type: string
    type: object
  database.FiveWhy:
    properties:
      answer:
        type: string
      why:
        type: string
    type: object
...
  rca.templateBody:
    properties:
      environmentFactor:
        type: string
      equipmentFactor:
        type: string
      fishbone:
        $ref: '#/definitions/database.Fishbone'
      fiveWhys:
        items:
          $ref: '#/definitions/database.FiveWhy'
        type: array
      humanFactor:
        type: string
      timeline:
        type: string
    type: object
  rca.upsertBody:
    properties:
      completed:
        type: boolean
      environmentFactor:
        type: string
      equipmentFactor:
        type: string
      fishbone:
        $ref: '#/definitions/database.Fishbone'
      fiveWhys:
        items:
          $ref: '#/definitions/database.FiveWhy'
        type: array
      humanFactor:
        type: string
      timeline:
        type: string
    type: object
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: K3 incident reporting and compliance API (MVP).
  title: Aegis API
  version: "1.0"
paths:
...
  /incidents/{id}/rca:
    get:
      parameters:
      - description: incident id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get RCA for incident
      tags:
      - rca
    put:
      consumes:
      - application/json
      parameters:
      - description: incident id
        in: path
        name: id
        required: true
        type: string
      - description: rca
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rca.upsertBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Create or update RCA
      tags:
      - rca
...
  /rca-templates/{category}:
    get:
      parameters:
      - description: incident category
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get RCA template by category
      tags:
      - rca
    put:
      consumes:
      - application/json
      parameters:
      - description: incident category
        in: path
        name: category
        required: true
        type: string
      - description: template
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rca.templateBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Save RCA template by category
      tags:
      - rca
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"